requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.4.1
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==20181003210633
idna==2.6
importlib-metadata==1.6.0
isort==4.3.21
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.4.3
luma.core==1.14.0
luma.led-matrix==1.4.1
mccabe==0.6.1
numpy==1.16.2
opencv-python==3.4.6.27
picamera==1.13
Pillow==7.1.2
pycrypto==2.6.1
pyftdi==0.51.0
pygame==1.9.4.post1
PyGObject==3.30.4
pylint==2.5.1
pyserial==3.4
python-apt==1.8.4.1
pyusb==1.0.2
pyxdg==0.25
requests==2.21.0
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
smbus2==0.3.0
spidev==3.4
ssh-import-id==5.7
tflite==2.1.0
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_armv7l.whl
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.1
virtualenv==20.0.18
wrapt==1.12.1
ws2812==1.0.0
zipp==3.1.0
